+++ migrations/003_convert_session_category_to_numeric.sql
@@ -33,5 +33,5 @@ SET category = CASE category
   END;
 
 ALTER TABLE tennis_sessions
-  MODIFY COLUMN category ENUM('1','2','3') NOT NULL DEFAULT '1' COMMENT '一级类型:1日常球局 2训练 3比赛',
-  MODIFY COLUMN sub_category ENUM('1','2','3','4') NOT NULL DEFAULT '2' COMMENT '二级类型:1单打/打单 2双打 3发球 4其他';
+  MODIFY COLUMN category SMALLINT NOT NULL DEFAULT 1 COMMENT '一级类型:1日常球局 2训练 3比赛',
+  MODIFY COLUMN sub_category SMALLINT NOT NULL DEFAULT 2 COMMENT '二级类型:1单打/打单 2双打 3发球 4其他';